refactor: file.goを分割
